Proper iterable map for executors
I've checked implementation of concurrent.futures for first time and was really confusing.
Actually I thought that it handles iterables much more gently but it creates a list underneath. With whole set of
iterables. It makes sense if you can fit everything in memory but for Streams case it won't work well at all. So
this is an implementation of map method suitable for my needs.